feat: moved shared components to components dir
