+++ src/components/Algorithms/SideMenu.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 // components
-import { Header } from '../../shared'
+import { Header } from '../shared'
 
 // utils
 import { ITableOfContent } from '../../utils'
