+++ src/components/Home/LegendContainer.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 // components
 import Legend from './Legend'
-import { Header } from '../../shared'
+import { Header } from '../shared'
 
 // utils
 import { BarStatus } from '../../utils'
