+++ src/components/shared/NavBar/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 // components
-import { Header } from '../'
+import { Header } from '..'
 
 interface IProps {
   children?: React.ReactNode
